pyproject.toml
[tool.poetry]
name = "devgyurak-log-backend"
version = "0.1.0"
description = ""
authors = ["devgyurak <devgyurak@gmail.com>"]
readme = "README.md"
packages = [{include = "devgyurak"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
